+++ Assets/Scripts/Interactables/Heart.cs
@@ -2,8 +2,8 @@
 {
     public override void Interact(PlayerInteraction player)
     {
-        int health = player.player.GetComponent<StatsManager>().currentHP;
-        int maxHealth = player.player.GetComponent<StatsManager>().maxHealth.value;
+        float health = player.player.GetComponent<StatsManager>().currentHP;
+        float maxHealth = player.player.GetComponent<StatsManager>().maxHealth.value;
 
         if (health < maxHealth)
         {
